// pharmacy-ui/src/app/services/models/user-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Roles } from '../models/roles';
export interface UserResponse {
  createdDate?: string;
  customerId?: number;
  dateOfBirth?: string;
  email?: string;
  firstName?: string;
  lastModifiedDate?: string;
  lastName?: string;
  location?: string;
  password?: string;
  phone?: string;
  roles?: Array<Roles>;
}